[SYCL 2020] Make array size a size_t

+++ SYCLStream.cpp
@@ -15,7 +15,7 @@ std::vector<sycl::device> devices;
 void getDeviceList(void);
 
 template <class T>
-SYCLStream<T>::SYCLStream(const int ARRAY_SIZE, const int device_index)
+SYCLStream<T>::SYCLStream(const size_t ARRAY_SIZE, const int device_index)
 : array_size {ARRAY_SIZE},
   d_a {ARRAY_SIZE},
   d_b {ARRAY_SIZE},
